extract txwrite
