+++ ydb/core/persqueue/read_balancer.cpp
@@ -1,5 +1,5 @@
 #include "read_balancer.h"
-#include "read_balancer_schema.h"
+#include "read_balancer_txwrite.h"
 
 #include <ydb/core/persqueue/events/internal.h>
 #include <ydb/core/protos/counters_pq.pb.h>
@@ -12,9 +12,6 @@
 namespace NKikimr {
 namespace NPQ {
 
-
-using namespace NTabletFlatExecutor;
-using namespace NPQRBPrivate;
 
 static constexpr TDuration ACL_SUCCESS_RETRY_TIMEOUT = TDuration::Seconds(30);
 static constexpr TDuration ACL_ERROR_RETRY_TIMEOUT = TDuration::Seconds(5);
@@ -147,61 +144,6 @@ void TPersQueueReadBalancer::TTxInit::Complete(const TActorContext& ctx) {
         Self->InitDone(ctx);
 }
 
-bool TPersQueueReadBalancer::TTxWrite::Execute(TTransactionContext& txc, const TActorContext&) {
-    NIceDb::TNiceDb db(txc.DB);
-    TString config;
-    bool res = Self->TabletConfig.SerializeToString(&config);
-    Y_ABORT_UNLESS(res);
-    db.Table<Schema::Data>().Key(1).Update(
-        NIceDb::TUpdate<Schema::Data::PathId>(Self->PathId),
-        NIceDb::TUpdate<Schema::Data::Topic>(Self->Topic),
-        NIceDb::TUpdate<Schema::Data::Path>(Self->Path),
-        NIceDb::TUpdate<Schema::Data::Version>(Self->Version),
-        NIceDb::TUpdate<Schema::Data::MaxPartsPerTablet>(Self->MaxPartsPerTablet),
-        NIceDb::TUpdate<Schema::Data::SchemeShardId>(Self->SchemeShardId),
-        NIceDb::TUpdate<Schema::Data::NextPartitionId>(Self->NextPartitionId),
-        NIceDb::TUpdate<Schema::Data::Config>(config),
-        NIceDb::TUpdate<Schema::Data::SubDomainPathId>(Self->SubDomainPathId ? Self->SubDomainPathId->LocalPathId : 0));
-    for (auto& p : DeletedPartitions) {
-        db.Table<Schema::Partitions>().Key(p).Delete();
-    }
-    for (auto& p : NewPartitions) {
-        db.Table<Schema::Partitions>().Key(p.PartitionId).Update(
-            NIceDb::TUpdate<Schema::Partitions::TabletId>(p.TabletId)
-        );
-    }
-    for (auto & p : NewGroups) {
-        db.Table<Schema::Groups>().Key(p.first, p.second).Update();
-    }
-    for (auto& p : NewTablets) {
-        db.Table<Schema::Tablets>().Key(p.first).Update(
-            NIceDb::TUpdate<Schema::Tablets::Owner>(p.second.Owner),
-            NIceDb::TUpdate<Schema::Tablets::Idx>(p.second.Idx));
-    }
-    for (auto& p : ReallocatedTablets) {
-        db.Table<Schema::Tablets>().Key(p.first).Update(
-            NIceDb::TUpdate<Schema::Tablets::Owner>(p.second.Owner),
-            NIceDb::TUpdate<Schema::Tablets::Idx>(p.second.Idx));
-    }
-    return true;
-}
-
-void TPersQueueReadBalancer::TTxWrite::Complete(const TActorContext &ctx) {
-    for (auto& actor : Self->WaitingResponse) {
-        THolder<TEvPersQueue::TEvUpdateConfigResponse> res{new TEvPersQueue::TEvUpdateConfigResponse};
-        res->Record.SetStatus(NKikimrPQ::OK);
-        res->Record.SetTxId(Self->TxId);
-        res->Record.SetOrigin(Self->TabletID());
-        ctx.Send(actor, res.Release());
-    }
-    Self->WaitingResponse.clear();
-
-    Self->NoGroupsInBase = false;
-    if (!Self->Inited) {
-        Self->Inited = true;
-        Self->InitDone(ctx);
-    }
-}
 
 struct TPersQueueReadBalancer::TTxWritePartitionStats : public ITransaction {
     TPersQueueReadBalancer * const Self;
